+++ release-notes.md
@@ -187,3 +187,6 @@ Release v0.3.2 - fix undo insert bug, specify tempo on timeline
 * NEW: Maneuvers menu
 * NEW: Countermarch, Illinois, Column, and Texas Turn maneuvers, with "footprints" preview
 * FIX: Sped up chart printing by about 50%.  Still takes about 1.5s per chart.
+
+
+TODO: need a way to undo a column... right-to-left option? undo option? different tool?