+++ release-notes.md
@@ -188,5 +188,4 @@ Release v0.3.2 - fix undo insert bug, specify tempo on timeline
 * NEW: Countermarch, Illinois, Column, and Texas Turn maneuvers, with "footprints" preview
 * FIX: Sped up chart printing by about 50%.  Still takes about 1.5s per chart.
 
-
-TODO: need a way to undo a column... right-to-left option? undo option? different tool?
+Release v0.3.3